REFACTOR: Remove unneeded code.

+++ SCSI/SSC/DensitySupport.cs
@@ -161,7 +161,7 @@ namespace DiscImageChef.Decoders.SCSI.SSC
                 if(descriptor.defaultDensity) sb.AppendLine("\tThis is the default density on the drive");
                 sb.AppendFormat("\tDensity has {0} bits per mm, with {1} tracks in a {2} mm width tape",
 #pragma warning disable IDE0004 // Remove Unnecessary Cast
-                                descriptor.bpmm, descriptor.tracks, (double)((double)descriptor.width / (double)10))
+                                descriptor.bpmm, descriptor.tracks, descriptor.width / (double)10)
                   .AppendLine();
 #pragma warning restore IDE0004 // Remove Unnecessary Cast
                 sb.AppendFormat("\tDensity maximum capacity is {0} megabytes", descriptor.capacity).AppendLine();
@@ -252,7 +252,7 @@ namespace DiscImageChef.Decoders.SCSI.SSC
 
                 sb.AppendFormat("\tMedium has a nominal length of {0} m in a {1} mm width tape",
 #pragma warning disable IDE0004 // Remove Unnecessary Cast
-                                descriptor.length, (double)((double)descriptor.width / (double)10)).AppendLine();
+                                descriptor.length, descriptor.width / (double)10).AppendLine();
 #pragma warning restore IDE0004 // Remove Unnecessary Cast
                 sb.AppendFormat("\tMedium description: {0}", descriptor.description).AppendLine();
                 sb.AppendLine();
